requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.1.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
selenium==4.15.0
webdriver-manager==4.0.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
lxml==4.9.3
fake-useragent==1.4.0
nest-asyncio==1.5.8
cloudscraper==1.2.71
pyparsing==3.1.1
ecdsa==0.18.0
pyasn1==0.5.1
attrs==23.1.0
outcome==1.3.0.post0
PySocks==1.7.1
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
wsproto==1.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.3
undetected-chromedriver==3.5.4